Fix BFloat16/Half mismatch in VoxCPM by using torch.bfloat16

+++ app.py
@@ -14,10 +14,10 @@ device = "cuda" if torch.cuda.is_available() else "cpu"
 model = VoxCPM.from_pretrained("openbmb/VoxCPM2", load_denoiser=False)
 if torch.cuda.is_available():
     try:
-        model.to(device="cuda", dtype=torch.float16)
+        model.to(device="cuda", dtype=torch.bfloat16)
     except AttributeError:
         if hasattr(model, 'tts_model') and model.tts_model is not None:
-            model.tts_model.to(device="cuda", dtype=torch.float16)
+            model.tts_model.to(device="cuda", dtype=torch.bfloat16)
         if hasattr(model, 'vocoder') and model.vocoder is not None:
             model.vocoder.to(device="cuda")
 sample_rate = getattr(getattr(model, "tts_model", None), "sample_rate", 16000)
